feat: add retries to yarn_command (IN-583) (#129)
* feat: add retries to yarn_command

* add echo to retry

* try with full retry logic

* try with full retry logic

+++ src/scripts/yarn/run_command.sh
@@ -4,9 +4,17 @@
 trap 'echo "fail detected"; touch /tmp/failure' ERR
 
 echo "Running command: ${COMMAND:?}"
-bash -c "${COMMAND:?}"
+for _ in $(seq 0 "${MAX_RETRIES:?}"); do
+    if bash -c "${COMMAND?}"; then
+        if (( ${SHOULD_REMOVE_LOCKFILE?} )); then
+            echo "Removing lock file ${LOCK_FILE?}" 
+            rm -rf "${LOCK_FILE}"
+        fi
+        exit 0
+    fi
+    sleep "${SLEEP_TIME:?}"
+    echo "Retrying command: ${COMMAND?}"
+done
 
-if (( ${SHOULD_REMOVE_LOCKFILE?} )); then
-    echo "Removing lock file ${LOCK_FILE?}" 
-    rm -rf "${LOCK_FILE}"
-fi
+echo "failed: ${COMMAND?}" >&2
+exit 1
